+++ packages/@codotype/core/src/DataPreview.ts
@@ -1,12 +1,13 @@
 export enum DataPreviewConstraintType {
   exists = "exists",
   contains = "contains",
-  equals = "equals"
+  equals = "equals",
 }
 
-export enum DataPreviewAction {
+export enum DataPreviewActionType {
   literal = "literal",
-  stringTemplate = "stringTemplate"
+  block = "block",
+  stringTemplate = "stringTemplate",
 }
 
 /**
@@ -17,32 +18,23 @@ export enum DataPreviewLayoutVariant {
   CODE_DARK = "CODE_DARK",
   CODE_LIGHT = "CODE_LIGHT",
   BADGE_LIGHT = "BADGE_LIGHT",
-  BADGE_DARK = "BADGE_DARK"
+  BADGE_DARK = "BADGE_DARK",
 }
 
-export interface DataPreviewRule {
-  // Condition
-  sourceProperty: string;
-  operation: DataPreviewConstraintType;
+export interface DataPreviewConstraint {
+  dataProperty: string;
+  type: DataPreviewConstraintType;
   value: string;
-  // # Action
-  action: {
-    type: DataPreviewAction;
-  };
+}
+
+export interface DataPreviewAction {
+  type: DataPreviewActionType;
   template: string;
 }
 
-// TODO - use DataPreviewRuleV2?
-export interface DataPreviewRuleV2 {
-  constraint: {
-    sourceProperty: string;
-    type: DataPreviewConstraintType;
-    value: string;
-  };
-  action: {
-    type: DataPreviewAction;
-    template: string;
-  };
+export interface DataPreviewRule {
+  constraint: DataPreviewConstraint;
+  action: DataPreviewAction;
 }
 
 export interface DataPreview {